// src/interfaces/i5sing/I5singSongUrl.ts
export interface I5singSongUrl {
    songid: string;
    songtype: string;
    squrl: string;
    squrlmd5: string;
    sqsize: string;
    sqext: string;
    hqurl: string;
    hqurlmd5: string;
    hqext: string;
    hqsize: string;
    lqurl: string;
    lqurlmd5: string;
    lqsize: string;
    lqext: string;
}